Bug#27831843: Encapsulate the reference count in TABLE_SHARE
Encapsulate the reference count in TABLE_SHARE to prevent callers from
modifying it in unintended ways. Add asserts to verify that the
reference count is only accessed by callers that hold the required
mutex, and to verify that the reference count doesn't go below zero.

Make some debug code in TABLE::init() take the required mutex before
it accesses the reference count.

Update the table_cache-t unit test to maintain the reference count
correctly, instead of setting it to a huge value to prevent it from
going down to zero.

Change-Id: I4fc33a2337f9661ca0fd07950de10a15cf085bd4