Bug#27297681 : Add error handling to ClusterTransactionImpl.scanIndex()
The scanIndex() function in ClusterTransactionImpl class, internally
calls ndbTransaction's scanIndex() to create the NdbIndexScanOperation.
The ndbTransaction's scanIndex() returns null if there is any error. But
the scanIndex() of ClusterTransactionImpl never checks this returned
operation and due to this, the ClusterJ app crashes with a
NullPointerException. This patch fixes that by checking the
NdbIndexScanOperation before returning it to the caller.