Bug#27490442 - MTR TESTS WITH ENCRYPT-PASSWORD FAILING ON COPY-BACK-AND-APPLY-LOG OF TTS BACKUP
Cause:
MEB does not support TTS backup of TDE tables. But there is a
possibility of TDE tables' redos being present in the redo log of the
backup.  During redo-log scan(apply-log operation)on such a backup it
tries to get the master-key assuming the redo will be applied on a TDE
table. As MEB has not backed up the KEF file, the call to return the
master key fails.

Fix:
MEB should skip parsing the redo logs of the tables not present in the
backup