Bug #25957586 BINLOGGING OF GENERATED NON STORED COLUMNS FAILS
A mysqld crash occurs when creating and using an NDB table with
non-stored generated columns. The crash only occurs with the
product compiled with asserts enabled(debug compile or release
without -DNBDEBUG). A mysqld hits this crash only if it is
configured to write binlogs.

The binlog injector can handle events which contain non-stored
generated columns, but it requires at least one stored column to
be present. If there are no stored columns, the binlog injector
fails to pack the row to write to the binlog.

Modified the binlog injector to skip writing the event if the event
contains no stored columns. Retained the behaviour that if an event
does contain a stored column, all non-stored columns are skipped.