Bug#24621170: SHOW PROCESSLIST x plugin
For X plugin, multiple Srv_sessions can be attached to the same thd.
When opening/closing a Srv_session on a certain thd, the server needs to
associate a protocol. Since this is not synchronised, there can be a
moment when the pointer to the protocol is null.
This fix synchronises the setting/getting of the protocol to avoid this
kind of problems.