Bug #27522405	SET PERSIST NOT HONORING TIMESTAMPS CAUSING A GR INIT FAILURE
Problem: When binlog_checksum is persisted to a value NONE and other needed GR
specific variables are persisted, next server start should make GR start on
boot which is not the case.

Analysis: Although persisted variables are sorted out based on timestamp.
Internally in code we maintain 2 separate list, 1 for server variables and 1
for plugin variables. In first iteration we set all persisted server variables
and in second iteration we set all plugin variables(during plugin load). Since
binlog_checksum is a dynamic variable it is set to its persisted value very
late in server startup phase. GR plugin variables although defined as dynamic
variables they are set even before GR plugin is loaded. Thus we see that GR
does not start on boot even when binlog_checksum is persisted before GR
variables. To be precise GR is loaded before the server loads the value for
binlog_checksum.

Fix: Fix is to introduce a flag PERSIST_AS_READ_ONLY which when set for server
variables will make the variable a pseudo read only variable. For any server
variable when this flag is set this variable is written to
"mysql_server_static_options" which will make variable to be set at command line
during server start.