Bug #24674276 MYSQLBINLOG -R --HEXDUMP CRASHES FOR INTVAR,                   USER_VAR, OR RAND EVENTS
This patch increases the test coverage of mysqlbinlog for
Intvar, User_var and Rand events by adding a testcase
covering the scenarios mentioned in the bug report.

Adding the testcase since the above mentioned bug is already
fixed by the patch to the Bug#26878022 and there was no
testcase covering the scenarios mentioned in the bug report.