WL#11453: [03/14] Item_func_coordinate_observer
The one-parameter ST_LATITUDE, ST_LONGITUDE, ST_X and ST_Y functions are
all variations of the same function. The only difference is in which
coordinate value is read, and if latitude/longitude conversion
happens. Therefore, add an abstract superclass for all functions that
return a coordinate value from a point.

Change-Id: I6178fcfafdc0dbe54018d9ff48b0f5d1089bb959