WL#11453: [08/14] gis::write_geometry
Add a geometry visitor, gis::Wkb_visitor, that produces a little-endian
WKB representation of a geometry.

Add function gis::write_geometry to produce a on-disk representation of
a geometry.

Rename sql/gis/wkb_parser.{h,cc} to sql/gis/wkb.{h,cc} to reflect that
it now contains functions to both read and write WKB (and the on-disk
format, which is derived from WKB).

Change-Id: I4ac2d91e517f50b5c5a36e19ec262253719d2e81