Bug #27522732 ADAPT NDBCLUSTER TO ENHANCED CHARSET SEMANTICS
The previous patches completely eliminated the usage of string
transformation for compare. What is left is string xfrm
where the result is later used for generating a hash value.

Naming of the xfrm functions is changed to reflect that
they now are for hash usage only. Also added comments about
this.

Added the debug hook 'verify_hash_only_usage' to NdbSqlUtil::strnxfrm_hash.
When active, the hash-transformed string will be all-zero, which will
be a legal hash transformation. Will create a lot of hash collisjons though,
which will be inneficient, but allowed for hash usage.
for compare however, this will create false equalities, which
will cause test failures. Thus, such missuse will be detected.