BUG#27081073 RENAMING A COLUMN BREAKS REPLICATION FROM 5.7 TO 8.0 BECAUSE OF IMPL. COLLATION
Problem and analysis:
Cross-version replication between a 5.7 master and a 8.0 slave may fail
because the default collation for utf8mb4 was changed in 8.0 and the
slave is not aware that it is replicating data with a different
collation.

Fix:
This patch replicates the value of the status variable
default_collation_for_utf8mb4 so that a slave is able to correctly
process data that originated from a master with a different default
collation for utf8mb4.
If the variable is not replicated, the slave assumes that it is
replicating an event from an older master and the variable is set to
5.7- default collation for utf8mb4 value, utf8mb4_general_ci.