BUG#27493720: INPLACE ALTER TABLE FAILS ON PARTICIPANT AFTER FIX FOR BUG#25915132
Problem: Cloning a dd::Table object only worked correctly if
sub-objects had been assigned a valid dd::Object_id. As this is done
when the object is stored, it meant that unstored object could not be
cloned. Root cause was that the Object_id was used to establish
cross-references inside the dd::Table object-tree when cloning.

Solution: Use ordinal position, rather than dd::Object_id when
establishing cross-references during clone. Using the ordinal position
provides a speedup compared to the linear search through a Collection
and comparing names. Comparing names would work even for objects which
have not been stored and thus do not have valid Object_ids.

For cloning we assume that all sub-objects in the source have correct
ordinal positions matching their current place in the Collection. It
is also assumed that the Column collection has been correctly cloned
before cloning the Index_elements, and the Index collection has been
correctly cloned before cloning the Partion_indexes. This was always
the case as the get_column() and get_index() apis used previously also
depended on the collections being cloned in the right order.

Change-Id: I56a0c19f2e5e54fbfdc899ef75072abdb4e5b1e9