Bug#27401817: ENFORCE THAT ALL MEMBERS DO HAVE THE SAME LOWER_CASE_TABLE_NAMES VALUE
Description:
Group Replication conflict detection does use schema and table names
as part of the Primary Key Equivalent (PKE) in order to detect and
disallow conflicting transactions.
The value of lower_case_table_names option does change how schema
and tables names are stored and externalized, which depending on the
configuration value may persist a T1 as t1. This difference on a
group can cause inconsistencies:
Example:
  M1: lower_case_table_names=0
  M2: lower_case_table_names=0
  M3: lower_case_table_names=1

  CREATE TABLE T1;
  would be stored as T1 on M1 and M2, but as t1 on M3.

Resolution:
When a server tries to join a group, its lower_case_table_names
value must be compared with the group one, if the value is different
the server must not be allowed to join.