Bug#26881703 INCORRECT BEHAVIOUR OCCURS WHEN OPEN TABLE IS                        EXPANDING VIEW WITH SAME NAME
Prepared stmt with CREATE TABLE ... SELECT leads to
unexpected server exit when a view exists with the same
name having a group by.

This is because the view definition is expanded during open
table and the resulting LEX object is attached to the select
LEX of the CREATE TABLE ... SELECT. fix_fields() for this
LEX will not be called as this is an invalid attachment for
CREATE TABLE flow and results in unexpected server exit
while traversing the item tree.

Fix: Do not expand the view definition if the open table is
invoked with TABLE_LIST::OPEN_FOR_CREATE set. Instead,
create a dummy view object to let the executor know of the
existence of a view.

Change-Id: Ic2fd60300018aaf17039427991c1912991025604