Bug #27460607: INCORRECT BEHAVIOR OF IODKU WHEN                INSERT SELECT's SOURCE TABLE IS EMPTY
Issue:
------
When the VALUES part of ON DUPLICATE KEY UPDATE contains a
BLOB value, it is handled using the
LEX::insert_update_values_map.

The issue in this bug is similar to the one in
Bug #25361251. LEX::insert_update_values_map is initialized
but not freed. This is because in an INSERT...SELECT
statement, if the SELECT part doesn't produce any rows,
write_record() isn't called.

Solution:
---------
As part of the fix for Bug #25361251, I added
LEX::clear_values_map() to write_record(). But a better
place to do the cleanup would be the end of the functions
Sql_cmd_insert_select::execute and  Sql_cmd_insert::execute.

This will make the design more similar to 8.0/trunk of having
LEX::clear_values_map() at the end of Sql_cmd_dml::execute().