Bug #27550943 MAKE THE DATA NODE LOG BUFFER SIZE CONFIGURABLE
In Bug #24748843, asynchronous IO was introduced in ndbd/ndbmtd.
This includes maintaining a log buffer in memory which stores logs
before it can be written to the disk.
There could be situations when the log buffer runs out of space if it
gets clogged with too many log messages; some log messages get lost due
to this.
Though this should not occur during normal operation, it is likely to
occur while debugging with huge amount of extra logging.

So, this patch adds capability to configure the log buffer size through
an option --logbuffer-size that can be given to ndbd/ndbmtd as a
command-line parameter.