BUG#27543150: MTR: READLINE() ON CLOSED FILEHANDLE FH AT MYSQL-TEST-RUN.PL               LINE 1179
Issue:
------
In PB2 runs, MTR generates "readline() on closed filehandle FH at
mysql-test-run.pl line 1179" warning on some platforms. This issue
started appearing after the fix for BUG#27361187.

Looks like the file containing the build id file locations created in
a child process, is not available for the parent process for some
reason. When MTR tries to open and read this file in the parent, the
above warning is generated.

Fix:
----
Create the file for storing build thread id file paths in parent
process itself, and use it at the end to cleanup all the build thread
id files.

Change-Id: I93618993839c648ef0c252bae3ce4d76d0a00283