Bug# 27191764 : AUTOTEST : ATRT DATABASE WRONGLY CONFIGURED TO USE CLUSTER
Description:
When ATRT generates processes configuration it assumes all MySQL
servers should have cluster connection options. While this might
be true for MySQL servers used in testing, this is not the case
for the ATRT database.

Update ATRT settings generation such that it no longer treats all
mysqld instances the same way and skips cluster and replication
features for atrt. Additionally, for a configuration specifying
just two replicated mysqld without cluster defined, update
settings such that ndb-connectstring is not generated and hence,
the servers don't try to connect to Cluster.