Bug #27576700 GIS: MYSQLD GOT SIGNAL 11 AT ../SQL/GIS/GEOMETRIES.CC:37
Problem: The server crashes during evaluation of ST_Crosses where at
least one operand is a geographic geometrycollection, at least one
operand is not of dimension 1, and the first operand contains fewer
points than the second operand.

While looping through the points, the index into the mulitipoint
containing the points from the second operand is used to access the
correspoinding multipoint for the first operand. When the index is
larger than the number of points in the first operand, the code
dereferences a null pointer.

Solution: Access elements of collections using indexes from the same
collection.

Change-Id: I05700c3ed55389338287962b14e41dc1bb90f3ae