Bug #27438963 ANALYZE TABLE CAN HOG CPU ON LARGE LOW CARDINALITY TABLES
As part of ANALYZE TABLE, secondary ordered indexes are scanned to
produce statistics about the distribution of key values, which is
useful when making query execution plans.

The statistical scanning is unusual in that aggregation is performed
intrinsically in the scan, aggregating information about the
distribution of keys in multiple rows into a single statistical
row returned by the scan.

The aggregation level depends on :
 - The size of the index (rows)
 - The desired size of the resulting statistics (resolution)
 - The cardinality of the index

When the aggregation level was high (many rows per statistic entry),
the statistical scan could run for a long time without taking a
break.

Running for a long time without a break is not permitted in the
Ndb data nodes, as it can lead to latency problems, throughput
stalls, watchdog warnings, heartbeat failures and node failures.

The statistical scanning logic is modified to take regular breaks
while aggregating rows, independently of the aggregation level.

This will avoid the negative consequences above, and should ensure
that ANALYZE TABLE has minimal impact on ongoing transactions.