Bug #14198580 --PRINT_DATA NOT PRINTING TRAILING 0S ON LONGVARBINARY FIELD
The values of columns of type BINARY, VARBINARY, LONGVARBINARY are printed differently
in mysql client and in ndb_restore when printed using hex format.
In the mysql client, trailing zeroes are printed whereas in ndb_restore, they are not.
This patch fixes this inconsistency by printing the trailing zeroes in ndb_restore as well.

Also re-record relevant  MTR result files since the trailing zeroes will now
be printed.