WL#11722 Reorganised some functions in Dblqh to make code more readable.
Increase the maximum scan direct count. Counted normal rows
reported back as 2 and those that only report that condition
was not met will only count as 1 row. Increased max scan
direct count to 16 and default from 6 to 16.

Made the amount of rows scanned in one real-time break adaptive.
If the load on JBB level is low we are allowed to execute a bit
more compared to otherwise.

More work on preparing code for next step of scan optimisations.