WL#11722 Step 3 --------------- Avoid translation from physical row id to row id and back again to physical row id when scanning in DBTUX. It makes sense to use row ids in DBACC and DBTUP since it is closely related to sending messages to other nodes and recovery that deals with logical rowids rather than physical rowids.
For DBTUX we only get a row id to be able to read the row in TUP and TUP
is faster to use physical row ids. So since TUX stores physical row ids
there is no use in converting it to a logical row id. Given that
LQH now uses different row id dependent on who is scanning we added a
number of comments on this fact. It is important for LQH to keep track
of what row ids are logical and which are physical.