Bug #27130881: MISLEADING ERRORS ON 'CREATE_COLLECTION_INDEX' COMMAND
Description:
============
Any error message coming from the plugin talks about 'member' or 
'constraint' but that is NOT what the user sees, so the error is useless
for any devapi implementation and we either translate the errors or
implement our own validation. None of this is desirable, if we align the
plugin with what the devapi describes above the validation would be 
centered on the plugin and everybody can use it right away.

Another issue is that any error on a 'member' does not include the
member position, which is not user friendly, including the member
position would improve usability.


Reviewed-by: Lukasz Kotula <lukasz.kotula@oracle.com>
Reviewed-by: Tomasz Stepniak <tomasz.s.stepniak@oracle.com>
RB:18924