BUG#27540354 - JAVASCRIPT-SPECIFIC ANNOTATION FOR 64 BIT INTEGERS IN .PROTO FILES
Description
===========
JavaScript doesn't have full support of 64 bit integers, protobuf 3.0 has an
annotation  (field option) which forces string as type of the field.

Fix
===
The annotation "jstype=JS_STRING" can't be applied to proto files, because MySQL
Server must support protobuf 2.5 and 2.6 in which the annotation doesn't exist.
For now the option is going to be applied as comment which can be manually
uncommented or using a script:

    optional uint64 v_unsigned_int = 3 /* ifdef PROTOBUF3 [jstype = JS_STRING] */;

RB: 18940
Reviewed-by: Tomasz Stepniak <tomasz.s.stepniak@oracle.com>
Reviewed-by: Grzegorz Szwarc <grzegorz.szwarc@oracle.com>