Bug #27612912: REMOVE MYISAM CACHING FUNCTIONS
MyISAM has a set of HA_EXTRA_*_CACHE flags that are used to set some explicit
madvise() flags, mostly on locked tables. Given that MyISAM is obsolete and
the OS most likely does a much better job of managing the cache without our
help anyway, we can just remove it. This removes a fair bit of complexity in
the upcoming iterator refactor.

Change-Id: Ife769347b2dfe9f4c319bf694d84900343923802