Bug#26710839 - ASSERTION FAILURE: ROW0MYSQL.CC:1076:PREBUILT->MAGIC_N == ROW_PREBUILT_ALLOCATED
This is because that there is one handler open for doing TRUNCATE in InnoDB,
this is fine before current drop+create mode, however, with the new mode,
InnoDB has to close the only open handler before dropping the old table.
After dropping the old table, if there is anything wrong during next steps,
InnoDB could not construct the m_prebuilt->table easily or properly,
then in later SE API calls, as long as they access the m_prebuilt->table,
it may hit all kinds of assertion because the table object could be broken
already.

So to fix this issue thoroughly, the best is server closes all opening
handlers before diving into InnoDB for TRUNCATE. Then InnoDB will open
and close the tables internally, without the effort to maintain the
row_prebuilt_t structure. In the patch, TRUNCATE operation will go into
the handler::create() API too, but with different SQL commands, then
InnoDB can identify the exact operation, and either create a table,
or truncate a table.

With this patch, some error messages for FK constraints becomes shorter
and cleaner. At the meantime, it's possible to truncate a corrupted table now.

RB: 17260
Reviewed by: Jimmy Yang <jimmy.yang@oracle.com>
Reviewed by: Sivert Sorumgard <sivert.sorumgaard@oracle.com>