Bug#27537572: THE REFORMATTING CAUSED A WOLFSSL COMPILATION ERROR ON WINDOWS
There is an #include guard making sure the SSL includes don't interfere
with out own. But the clang-format shuffled them, ordering them
alphabetically. Theoretically this should be possible, but atm it is
not, so moving them back to the correct order.
Adding clang-format off guard not to shuffle them back again

Tested compiling on Windows with both WolfSSL and OpenSSL