WL#11722: Step 14 ----------------- During a scan we use two bounds, the starting boundary. This is used once in scanFirst to find the first row in the range. There is no need to precompute this boundary since it is only used once in a scan. Similarly we make no special optimisations for now for scans for table statistics.
The boundary used to check if the next row is outside the boundary
is however used many times, most of the time at least once, but
often it can be used many times over and over again. So this
boundary condition we precompute the objects for and use them
multiple times.

This optimises the scanCheck method greatly.