BUG#27635398: MTR: REFACTOR THE CODE FOR '--ERROR' MYSQLTEST COMMAND
Description:
------------
The mysqltest command '--error' takes comma separated list of error codes
that the next statement is expected to return. BUG#27121105 requests for
similar feature for '--disable_warnings' and for '--enable_warnings'
command. But the existing code written for '--error' command is not optimal
and is difficult to reuse.

Fix:
----
Refactor the code for '--error' mysqltest command.

1. New class 'Error' to store the information about a error

2. New class 'Expected_errors' containing list of errors given to --error
   mysqltest command.

3. List of errors are now stored in std::vector<> instead of an array of
   structures.

4. Modified following methods to use class functions to fetch the
   information regarding errors.

   - do_get_errcodes()
   - handle_command_error()
   - handle_error()
   - handle_no_error()
   - match_expected_error()

5. Cleaned up unnecessary code.

Change-Id: Ic50a0434d1e8d6c7781ca99f6336088a66975f04