Bug#27566220 PFS STAGE REAPPEARING WITHOUT BEING SET
Before this fix,
spurious stages can be seen in table
performance_schema.events_stages_current.

The root cause is that some threads may terminate
with an open stage.
Then, later, a new thread is started,
which reuse the same memory for the thread instrumentation.

Function create_thread() in pfs_instr.cc does not
explicitly reset the current stage,
so a left over stage from another thread run could be seen.

The fix is to reset the current stage in create_thread()
when a new thread instrumentation is created.