BUG#27472087: ALTER TABLE HANGS IN "WAITING FOR TABLESPACE METADATA LOCK" STATE
Problem: MDLs are now acquired by SE on implicit tablespaces. But SE
does not consider locked tables mode, which is accessible only to
server.

Solution: Add hook function which allows SE to notfy server about the
MDL_tickets used for the old and new implicit tablespace names. Add code
in the server which releases the lock on the old name and changes the
duration for the lock on the new name, when the transaction commits.

Change-Id: I594f3d29db32c497d07e1b2fdd732e8df87ba310