Bug#27577704 - INNODB: ASSERTION FAILURE: HA_INNOPART.CC:3331:IB_TABLE->STAT_INITIALIZED Bug#27586419 - ASSERT: TABLE->STAT_INITIALIZED::DICT_STATS_ASSERT_INITIALIZED
These two bugs should be due to the same cause, which is a regression from
bug#26848711. In that bug, the dict_sys mutex was removed for
dict_table_close(). So it requires a per-table lock to prevent the race
between table->n_rec_count and table->stats_initialized. The problems are
in that patch, it only forced the table lock around table->acquire() in
ha_innobase::open(). This is far not enough. In fact, all acquiring tables
when the table is already in memory should acquire the table lock, because
in all those code pathes, the table->stats_initialized would be checked,
and the race should be prevented too.

The patch fixes all these two bugs, by introducing a new dict_table_t API
called acquire_with_lock(), which will lock() and unlock() when acquiring
the table. This API should will replace normal acquire() calls when the
table to be opened is in memory and there would be stats initialization
afterwards. Comments related are also modified a bit to clearly demonstrate
this point.

RB: 19020
Reviewed by: Jimmy Yang <jimmy.yang@oracle.com>