Bug#27282595 NDB_RESTORE WITH --PRINT-DATA IS CRASHING
Skipped the call to check_blobs if we are only printing data.
If don't restore meta-data then the staging table list is not
populated to find the BLOB tables needed by check_blobs.
Extended ndb_restore_print with a testcase that only prints data
for a table with a BLOB clolumn.