WL #11590: More flexible filesort [patch 9/10, optimistic sort key generation]
Enable optimistic sort key generation.

With this change, we no longer end a filesort chunk when it is pessimistically
out of space (ie., can no longer sustain a worst-case row). Instead, we simply
start writing the row, and end the chunk only if it actually needed more space
than the chunk had left (or if it was within a few bytes -- we can't always tell).

This enables us to fit more rows into each chunk (for some cases, quite
a lot of more rows, although the common case is much more modest), as most rows
are not worst-case; even more so with the newer Unicode collations. Also,
it significantly reduces the need for artificially limiting the worst-case
size of rows through the max_sort_length variable; it is now only used for
blobs of PAD SPACE collations, where we have to pad the sort lengths to the
maximum size and the maximum size is by definition large. This means users
can now ORDER BY things like JSON columns or large VARCHAR fields without
having to worry that they get wrong results as sort keys are truncated.

Note that this means that you can no longer ORDER BY <json> LIMIT <n>
(or similarly, on a blob) and get ordering by priority queue, since the
priority queue still is oriented around worst-case rows. Hopefully, this
is not a very common use case, and getting actually correct results should
more than weigh up for it.

In the process, fix so that blobs get variable-length sort keys for
NO PAD collations. (It would be possible to fix it in a separate commit
earlier, but not without rewriting tests that would only need to be
re-rewritten now, so it's easiest to do it in the same commit.)

The unit tests for the filesort buffer have been completely rewritten,
since the working of the buffer is so different now.

Change-Id: I4c118dd7820549d47fad95e7bea7686cc7df619d