Bug#27566937 SERVER STARTUP CRASH AFTER SHUTDOWN IF TABLES CREATED USING JUNK CHARACTER-DEBUG
Problem :
---------
DROP DATABASE failed if the tables were created using multi-byte character strings. The issue
happens because the server could not find the location of the files because of some directory
separator character in the table and tablespace name. Because of these characters, the server
incorrectly parsed the tablespace name to find the database directory and the files within it.

Solution :
----------
Iterate through the tablespace name backwards to find the directory separator characters and
compare the directory name against the file path. This is done until the server finds
a valid directory.

Reviewed by: Kevin Lewis (kevin.lewis@oracle.com)

RB 19057